requirements.txt
Flask==3.0.1
flask_login
pymongo
python-dotenv